Setup firestore user data listener
